clear query hostory
