+++ src/components/Login/RecentConnections.tsx
@@ -1,29 +1,13 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { useState } from 'react';
 import ReactTooltip from 'react-tooltip';
+import DeleteIcon from '../atoms/DeleteIcon';
 import Utils from '../utils/utils';
 
 interface RecentConnectionsProps {
   send: (v?: string) => void;
   setConnectionString: React.Dispatch<React.SetStateAction<string>>;
 }
-const DeleteIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="12"
-    height="12"
-    viewBox="0 0 12 12"
-    // fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="feather feather-x"
-  >
-    <line x1="9" y1="3" x2="3" y2="9" />
-    <line x1="3" y1="3" x2="9" y2="9" />
-  </svg>
-);
 
 const RecentConnections: React.FC<RecentConnectionsProps> = ({
   send,
