+++ src/components/atoms/DeleteIcon.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const DeleteIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="12"
+    height="12"
+    viewBox="0 0 12 12"
+    // fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    className="feather feather-x"
+  >
+    <line x1="9" y1="3" x2="3" y2="9" />
+    <line x1="3" y1="3" x2="9" y2="9" />
+  </svg>
+);
+export default DeleteIcon;
