+++ src/state/reducer.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-case-declarations */
 import DbSession from '../sessions/DbSession';
 
 export type HistoryType = {
@@ -16,12 +17,15 @@ export type AppStateActions =
   | { type: 'SET_SESSION'; payload: DbSession }
   | { type: 'SET_HISTORY'; payload: HistoryType[] }
   | { type: 'ADD_HISTORY'; payload: HistoryType }
+  | { type: 'REMOVE_HISTORY'; payload: number }
+  | { type: 'CLEAR_HISTORY' }
   | { type: 'test'; payload: DbSession };
 
 export function appReducer(
   state: AppState = { history: [] },
   action: AppStateActions
 ): AppState {
+  // console.log(action, state);
   switch (action.type) {
     case 'SET_SESSION':
       return { ...state, session: action.payload };
@@ -32,6 +36,18 @@ export function appReducer(
       return {
         ...state,
         history: [action.payload, ...(state?.history ?? [])],
+      };
+    case 'REMOVE_HISTORY':
+      const newHistory = [...(state.history ?? [])];
+      newHistory?.splice(action.payload, 1);
+      return {
+        ...state,
+        history: newHistory,
+      };
+    case 'CLEAR_HISTORY':
+      return {
+        ...state,
+        history: [],
       };
     default:
       return state;
